refactor(indexer/rpc): extract readContractWithBlockFallback to rpc/block-fallback.ts (PR-S7)
Move ~155 lines out of indexer-envio/src/rpc.ts (1625 → 1474 lines) into
the new indexer-envio/src/rpc/block-fallback.ts (172 lines):

- `BLOCK_NOT_AVAILABLE_RE`, `BLOCK_RETRY_DELAYS_MS`
- `BlockFallbackResult` type
- `_testHooks` (delayFn proxy + the rate-limit/log/reset hooks for
  rpcClient.test.ts) — moved alongside its only consumer
  (`readContractWithBlockFallback`)
- `readContractWithBlockFallback` core primitive — handles both
  rate-limit retries (with secondary-RPC fallback) and "block not
  available" retries (with `latest` fallback)

`rpc.ts` re-exports `readContractWithBlockFallback`, `_testHooks`, and
the `BlockFallbackResult` type so downstream callers (the in-file
fetchers + blockFallback.test.ts + rpcClient.test.ts via "../src/rpc")
keep working without import-path changes.

Pure extraction. Pinned by `blockFallback.test.ts` (54 tests) +
`rpcClient.test.ts` (24 tests via the re-exported `_testHooks`).

`pnpm test` reports 381 passing.
